Server/Plugin: Move OnDemandDict to helpers
The OnDemandDict could be used by different plugins.
